+++ Model/Request/Payment/Refund.php
@@ -113,10 +113,15 @@ class Refund extends AbstractPayment implements RequestInterface
             $order->getId()
         );
 
-		$authCode		= $orderPayment->getAdditionalInformation(Payment::TRANSACTION_AUTH_CODE_KEY);
-        $payment_method	= $orderPayment->getAdditionalInformation(Payment::TRANSACTION_EXTERNAL_PAYMENT_METHOD);
+//		$authCode		= $orderPayment->getAdditionalInformation(Payment::TRANSACTION_AUTH_CODE_KEY);
+		$sale_settle_params	= $orderPayment->getAdditionalInformation(Payment::SALE_SETTLE_PARAMS);
+		
+		$this->config->createLog($sale_settle_params, 'Refund sale_settle_params');
+		
+        $payment_method		= $orderPayment->getAdditionalInformation(Payment::TRANSACTION_EXTERNAL_PAYMENT_METHOD);
 
-        if (empty($authCode) && Payment::APM_METHOD_CC == $payment_method) {
+//        if (empty($authCode) && Payment::APM_METHOD_CC == $payment_method) {
+        if (empty($sale_settle_params['AuthCode']) && Payment::APM_METHOD_CC == $payment_method) {
             $msg = __('Transaction does not contain authorization code.');
             $this->config->createLog($msg);
             throw new PaymentException($msg);
@@ -127,7 +132,8 @@ class Refund extends AbstractPayment implements RequestInterface
             'currency'              => $order->getBaseCurrencyCode(),
             'amount'                => (float)$this->amount,
             'relatedTransactionId'    => $transactionId,
-            'authCode'              => is_null($authCode) ? '' : $authCode,
+//            'authCode'              => is_null($authCode) ? '' : $authCode,
+            'authCode'              => empty($sale_settle_params['AuthCode']) ? '' : $sale_settle_params['AuthCode'],
             'comment'               => '',
             'merchant_unique_id'    => $order->getIncrementId(),
             'urlDetails'            => [